+++ openaerostruct/tests/test_simple_rect_mphys_aero.py
@@ -94,7 +94,7 @@ class Test(unittest.TestCase):
                 dvs.add_output("cg", val=cg, units="m")
 
                 # Create mphys builder for aero solver
-                options = {"compressible": False}
+                options = {"compressible": False, "write_solution": False}
                 aero_builder = AeroBuilder(surfaces, options)
                 aero_builder.initialize(self.comm)
